warnings removed
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@3049 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSBuildCells.h
@@ -93,7 +93,7 @@ private:
     Cont _cellsCont;
 
     /// The Number of Cells in x- and y-Direction
-    int _xsize, _ysize;
+    size_t _xsize, _ysize;
 
 	/// The sizes of each cell
     SUMOReal _xcellsize, _ycellsize;
